Workaround for IDE error when placing App Engine Local Run menu item.
