+++ app-engine/java/resources/META-INF/app-engine-java.xml
@@ -35,11 +35,6 @@
     </extensionPoints>
 
     <actions>
-        <group id="GoogleCloudTools.AppEngine.General" class="com.google.cloud.tools.intellij.appengine.java.util.AppEngineGeneralMenuGroupAction"
-                popup="true">
-            <add-to-group group-id="GoogleCloudTools" anchor="first"/>
-        </group>
-
         <group id="GoogleCloudTools.AppEngine.Support">
             <group id="GoogleCloudTools.AddAppEngineFrameworkSupport"
                    class="com.google.cloud.tools.intellij.appengine.java.facet.AddAppEngineFrameworkSupportMenuGroupAction"
