+++ app-engine/java/resources/messages/AppEngineBundle.properties
@@ -194,7 +194,6 @@ appengine.run.settings.adminhost.label=Admin host:
 appengine.run.settings.adminport.label=Admin port:
 appengine.infopanel.no.region=Can't retrieve region for selected project.
 
-appengine.general.group.menu.text=App Engine
 appengine.add.facet.group.menu.text=Add App Engine support
 appengine.standard.facet.name=Google App Engine standard
 appengine.flexible.facet.name=Google App Engine flexible
