+++ google-cloud-core/resources/messages/CoreMessageBundle.properties
@@ -17,6 +17,7 @@
 
 cloud.code.top.group.menu.text=Cloud Tools
 cloud.tools.group.menu.text=Google Cloud Platform
+appengine.general.group.menu.text=App Engine
 
 cloud.project.selector.signin.message=Sign in with your Google account to list your Google Developers Console projects.
 cloud.project.selector.open.dialog.tooltip=Opens a dialog to select a Google Cloud project
