+++ google-cloud-core/src/com/google/cloud/tools/intellij/action/CloudToolsMenuGroupAction.java
@@ -14,18 +14,21 @@
  * limitations under the License.
  */
 
-package com.google.cloud.tools.intellij;
+package com.google.cloud.tools.intellij.action;
 
+import com.google.cloud.tools.intellij.GoogleCloudCoreIcons;
+import com.google.cloud.tools.intellij.GoogleCloudCoreMessageBundle;
 import com.intellij.openapi.actionSystem.DefaultActionGroup;
 
 /**
- * Top menu group for all plugin visible actions that are expressed with menu items. All K8s and
- * Google Cloud functionality is nested under this group. See plugin.xml for structure.
+ * Creates the tools menu top-level menu group for Cloud Tools. All relevant shortcuts are sub-menu
+ * items of this.
  */
-public class CloudCodeTopMenuGroupAction extends DefaultActionGroup {
-  public CloudCodeTopMenuGroupAction() {
+public class CloudToolsMenuGroupAction extends DefaultActionGroup {
+
+  public CloudToolsMenuGroupAction() {
     getTemplatePresentation()
-        .setText(GoogleCloudCoreMessageBundle.message("cloud.code.top.group.menu.text"));
+        .setText(GoogleCloudCoreMessageBundle.message("cloud.tools.group.menu.text"));
     getTemplatePresentation().setIcon(GoogleCloudCoreIcons.CLOUD);
   }
 }
